reorganize index route
There are gonna be two columns, which are more independent testable
units than a single monolith. So it's just better for everyone if
I change this now. Also change the tests to focus on ChangesColumn
instead of near-empty IndexHandler. (which they already did btw.)
